+++ PSNovusTools/Private/Close-SSHSession.ps1
@@ -1,13 +1,13 @@
 function Close-SSHSession([SSH.SshSession]$session) {
 
     try{
-        $result = Remove-SSHSession -SessionId $session.SessionId 
+        $sshresult = Remove-SSHSession -SessionId $session.SessionId 
       }
       catch {
-          Write-Error "Failed to close SSH Session."
+          Write-Error "Failed to close SSH Session. Error: $_"
       
           
       }
 
-      return $result   
+      return $sshresult   
 }