+++ PSNovusTools/Private/Test-SSHSession.ps1
@@ -1,0 +1,4 @@
+function Test-SSHSession([SSH.SshSession]$session) {
+
+     Write-Verbose "Session: $session"
+}